chore: move "GraphQLPayloadContext" to "typeUtils"
