--- src/context/types.ts
@@ -1,3 +1,0 @@
-import { ResponseTransformer } from '../response'
-
-export type GraphQLPayloadContext<T> = (payload: T) => ResponseTransformer
