+++ src/handlers/GraphQLHandler.ts
@@ -7,7 +7,7 @@ import { fetch } from '../context/fetch'
 import { data } from '../context/data'
 import { extensions } from '../context/extensions'
 import { errors } from '../context/errors'
-import { GraphQLPayloadContext } from '../context/types'
+import { GraphQLPayloadContext } from '../typeUtils'
 import { cookie } from '../context/cookie'
 import {
   MockedRequest,
@@ -35,7 +35,7 @@ export type GraphQLHandlerNameSelector = DocumentNode | RegExp | string
 // GraphQL related context should contain utility functions
 // useful for GraphQL. Functions like `xml()` bear no value
 // in the GraphQL universe.
-export type GraphQLContext<QueryType> = {
+export type GraphQLContext<QueryType extends Record<string, unknown>> = {
   set: typeof set
   status: typeof status
   delay: typeof delay
